+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -9,7 +9,8 @@ The Azure MCP Server updates automatically by default whenever a new release com
 ### Breaking Changes
 
 - Changed the following commands to change / remove available options
-  - Storage account creation removed the ability to configure `enable-https-traffic-only` (always true now), `allow-blob-public-access` (always false now), and `kind` (always StroageV2 now).
+  - Storage account create removed the ability to configure `enable-https-traffic-only` (always true now), `allow-blob-public-access` (always false now), and `kind` (always StroageV2 now).
+  - Storage blob container create removed the ability to configure `blob-container-public-access` (always false now).
 
 ### Bugs Fixed